fix hw3,4

+++ src/main/java/com/ifmo/lesson2/Factorial.java
@@ -18,7 +18,7 @@ public class Factorial {
     }
 
     public static long factorial(int n) {
-        int result = 1;
+        long result = 1;
         for (int i = 1; i <=n; i ++){
             result = result*i;
         }
